IPFS Calls from User Wallet #270
